update carNumberInput v1
