+++ src/components/carKeyboard/index.js
@@ -1,0 +1,3 @@
+import CarInput from './carInput.vue';
+
+export default CarInput;
