+++ src/components/carKeyboard/keyboardData.js
@@ -1,0 +1,107 @@
+export const provinces = [
+  [
+    { code: 31908, value: '粤' },
+    { code: 20140, value: '京' },
+    { code: 27941, value: '津' },
+    { code: 27818, value: '沪' },
+    { code: 28189, value: '渝' },
+    { code: 20864, value: '冀' },
+    { code: 35947, value: '豫' },
+    { code: 20113, value: '云' },
+    { code: 36797, value: '辽' },
+  ],
+  [
+    { code: 40657, value: '黑' },
+    { code: 28248, value: '湘' },
+    { code: 30358, value: '皖' },
+    { code: 40065, value: '鲁' },
+    { code: 26032, value: '新' },
+    { code: 33487, value: '苏' },
+    { code: 27993, value: '浙' },
+  ],
+  [
+    { code: 36195, value: '赣' },
+    { code: 37122, value: '鄂' },
+    { code: 26690, value: '桂' },
+    { code: 29976, value: '甘' },
+    { code: 26187, value: '晋' },
+    { code: 33945, value: '蒙' },
+    { code: 38485, value: '陕' },
+  ],
+  [
+    { code: 21513, value: '吉' },
+    { code: 38397, value: '闽' },
+    { code: 36149, value: '贵' },
+    { code: 38738, value: '青' },
+    { code: 34255, value: '藏' },
+    { code: 24029, value: '川' },
+    { code: 23425, value: '宁' },
+    { code: 29756, value: '琼' },
+    // { code: 20351, value: '使' },
+  ],
+];
+
+export const numbers = [
+  { code: 49, value: '1' },
+  { code: 50, value: '2' },
+  { code: 51, value: '3' },
+  { code: 52, value: '4' },
+  { code: 53, value: '5' },
+  { code: 54, value: '6' },
+  { code: 55, value: '7' },
+  { code: 56, value: '8' },
+  { code: 57, value: '9' },
+  { code: 48, value: '0' },
+];
+
+export const englishKeys = [
+  [
+    { code: 81, value: 'Q' },
+    { code: 87, value: 'W' },
+    { code: 69, value: 'E' },
+    { code: 82, value: 'R' },
+    { code: 84, value: 'T' },
+    { code: 89, value: 'Y' },
+    { code: 85, value: 'U' },
+    { code: 73, value: 'I' },
+    { code: 79, value: 'O' },
+    { code: 80, value: 'P' },
+  ],
+  [
+    { code: 65, value: 'A' },
+    { code: 83, value: 'S' },
+    { code: 68, value: 'D' },
+    { code: 70, value: 'F' },
+    { code: 71, value: 'G' },
+    { code: 72, value: 'H' },
+    { code: 74, value: 'J' },
+    { code: 75, value: 'K' },
+    { code: 76, value: 'L' },
+  ],
+  [
+    { code: 90, value: 'Z' },
+    { code: 88, value: 'X' },
+    { code: 67, value: 'C' },
+    { code: 86, value: 'V' },
+    { code: 66, value: 'B' },
+    { code: 78, value: 'N' },
+    { code: 77, value: 'M' },
+  ],
+];
+
+export const specialKeys = [
+  { code: 25346, value: '挂' },
+  { code: 23398, value: '学' },
+  { code: 35686, value: '警' },
+  { code: 28207, value: '港' },
+  { code: 28595, value: '澳' },
+  // { code: 20351, value: '使' },
+  { code: 39046, value: '领' },
+];
+
+export default {
+  provinces,
+  numbers,
+  englishKeys,
+  specialKeys,
+};
